+++ rqpy/plotting/__init__.py
@@ -1,4 +1,4 @@
-from ._core import *
-from ._energy_cal import *
-from ._fitting import *
-from ._iv_didv_tools import *
+from ._core_plotting import *
+from ._energy_cal_plotting import *
+from ._fitting_plotting import *
+from ._iv_didv_tools_plotting import *
